[backend/frontend] Update a simulation loose the automated trigger (#2914)
