+++ openbas-api/src/main/java/io/openbas/rest/exercise/form/ExerciseInput.java
@@ -1,16 +1,12 @@
 package io.openbas.rest.exercise.form;
 
-import static io.openbas.config.AppConfig.*;
-import static lombok.AccessLevel.NONE;
+import static io.openbas.config.AppConfig.EMAIL_FORMAT;
+import static io.openbas.config.AppConfig.MANDATORY_MESSAGE;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.annotation.Nullable;
 import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.NotBlank;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.Data;
@@ -45,12 +41,6 @@ public class ExerciseInput {
   @JsonProperty("exercise_description")
   private String description;
 
-  @Schema(nullable = true)
-  @JsonProperty("exercise_start_date")
-  @FutureOrPresent(message = NOW_FUTURE_MESSAGE)
-  @Getter(NONE)
-  private Instant start;
-
   @JsonProperty("exercise_tags")
   private List<String> tagIds = new ArrayList<>();
 
@@ -66,8 +56,4 @@ public class ExerciseInput {
 
   @JsonProperty("exercise_message_footer")
   private String footer;
-
-  public Instant getStart() {
-    return start != null ? start.truncatedTo(ChronoUnit.MINUTES) : null;
-  }
 }
